PPR API version upgrade step 1.
