+++ ppr-api/requirements.txt
@@ -1,62 +1,64 @@
-Babel==2.9.1
-Flask-Babel==2.0.0
-Flask-SQLAlchemy==2.5.1
-Flask==1.1.4
-Jinja2==2.11.3
-MarkupSafe==2.0.1
-PyYAML==6.0
-SQLAlchemy==1.4.31
-Werkzeug==1.0.1
+Babel==2.12.1
+Flask-Cors==3.0.10
+Flask-Migrate==4.0.4
+Flask-Moment==1.0.5
+Flask-SQLAlchemy==3.0.3
+Flask==2.2.3
+Jinja2==3.1.2
+Mako==1.2.4
+MarkupSafe==2.1.2
+SQLAlchemy==2.0.6
+Werkzeug==2.2.3
 aniso8601==9.0.1
-attrs==21.4.0
-blinker==1.4
-cachelib==0.6.0
-cachetools==4.2.4
-certifi==2021.10.8
-charset-normalizer==2.0.10
-click==7.1.2
+attrs==22.2.0
+blinker==1.5
+cachelib==0.10.2
+cachetools==5.3.0
+certifi==2022.12.7
+charset-normalizer==3.1.0
+click==8.1.3
 datedelta==1.4
-ecdsa==0.17.0
-expiringdict==1.1.4
+ecdsa==0.18.0
+expiringdict==1.2.2
+flask-babel==3.0.1
 flask-jwt-oidc==0.3.0
-flask-restx==0.5.1
-google-api-core==2.4.0
-google-auth-oauthlib==0.4.6
-google-auth==2.3.3
-google-cloud-pubsub==2.9.0
-google-cloud-storage==2.3.0
-googleapis-common-protos==1.54.0
-grpc-google-iam-v1==0.12.3
-grpcio-status==1.43.0
-grpcio==1.43.0
+google-api-core==2.11.0
+google-auth-oauthlib==1.0.0
+google-auth==2.16.2
+google-cloud-core==2.3.2
+google-cloud-pubsub==2.15.0
+google-cloud-storage==2.7.0
+googleapis-common-protos==1.58.0
+grpc-google-iam-v1==0.12.6
+grpcio-status==1.51.3
+grpcio==1.60.1
 gunicorn==20.1.0
-idna==3.3
-itsdangerous==1.1.0
+idna==3.4
+itsdangerous==2.1.2
 jsonschema==4.4.0
-launchdarkly-server-sdk==7.3.0
-libcst==0.4.0
+launchdarkly-server-sdk==8.1.1
+libcst==1.1.0
 mypy-extensions==0.4.3
-oauthlib==3.1.1
-proto-plus==1.19.8
-protobuf==3.19.3
-psycopg2-binary==2.9.3
+oauthlib==3.2.2
+proto-plus==1.22.2
+protobuf==4.22.1
+psycopg2-binary==2.9.9
 pyRFC3339==1.1
 pyasn1-modules==0.2.8
 pyasn1==0.4.8
 pycountry==22.1.10
-PyPDF2==2.10.2
-pyrsistent==0.18.1
-python-dotenv==0.19.2
+PyPDF2==3.0.1
+pyrsistent==0.19.3
+python-dotenv==1.0.0
 python-jose==3.3.0
-pytz==2021.3
-requests-oauthlib==1.3.0
-requests==2.27.1
-rsa==4.8
-semver==2.13.0
+pytz==2022.7.1
+requests-oauthlib==1.3.1
+requests==2.28.2
+rsa==4.9
+semver==2.13.0 
 sentry-sdk==1.20.0
 six==1.16.0
 strict-rfc3339==0.7
 typing-inspect==0.7.1
-typing_extensions==4.0.1
-urllib3==1.26.11
-git+https://github.com/bcgov/registry-schemas.git@1.7.3#egg=registry_schemas
+typing_extensions==4.9.0
+urllib3==1.26.15
