+++ ppr-api/setup.cfg
@@ -17,7 +17,7 @@ keywords =
 
 [options]
 zip_safe = True
-python_requires = >=3.6
+python_requires = >=3.9
 include_package_data = True
 packages = find:
 
@@ -68,11 +68,11 @@ good-names=
 
 [pylint]
 ignore=migrations,tests
-max_line_length=120
+max-line-length=120
 notes=FIXME,XXX,TODO
 ignored-modules=flask_sqlalchemy,sqlalchemy,SQLAlchemy,alembic,scoped_session
 ignored-classes=scoped_session
-disable=C0301,W0511,W0703,R0801,R0902,R0401
+disable=C0301,E1136,W0511,W0703,R0801,R0902,R0401,R0914,R0916,C0209,W0707
 
 [isort]
 line_length = 120
@@ -87,7 +87,7 @@ testpaths = tests
 addopts = --verbose
     --strict
     -p no:warnings
-    --cov=src --cov-report html:htmlcov --cov-report xml:coverage.xml
+#    --cov=src --cov-report html:htmlcov --cov-report xml:coverage.xml
 python_files = tests/*/test*.py
 norecursedirs = .git .tox venv* requirements* build
 log_cli = true
@@ -101,10 +101,11 @@ markers =
 [coverage:run]
 branch = True
 source =
-    src/legal_api
+    src/mhr_api
 omit =
-    src/legal_api/wsgi.py
-    src/legal_api/gunicorn_config.py
+    src/ppr_api/wsgi.py
+    src/ppr_api/gunicorn_config.py
+    src/database
 
 [report:run]
 exclude_lines =
