+++ ppr-api/src/ppr_api/common/enum.py
@@ -27,8 +27,8 @@ class BaseMeta(EnumMeta):
             self(other)  # pylint: disable=no-value-for-parameter
         except ValueError:
             return False
-        else:
-            return True
+
+        return True
 
 
 class BaseEnum(str, Enum, metaclass=BaseMeta):
@@ -39,6 +39,14 @@ class BaseEnum(str, Enum, metaclass=BaseMeta):
         """Return the enum by value."""
         for enum_value in cls:
             if enum_value.value == value:
+                return enum_value
+        return None
+
+    @classmethod
+    def get_enum_by_name(cls, value: str) -> Optional[str]:
+        """Return the enum by value."""
+        for enum_value in cls:
+            if enum_value.name == value:
                 return enum_value
         return None
