+++ ppr-api/src/ppr_api/reports/v2/report.py
@@ -18,12 +18,11 @@ import markupsafe
 import pycountry
 import requests
 from flask import current_app, jsonify
-
+from ppr_api.callback.auth.token_service import GoogleStorageTokenService
 from ppr_api.exceptions import ResourceErrorCodes
 from ppr_api.models import utils as model_utils
 from ppr_api.reports.v2 import report_utils
-from ppr_api.reports.v2.report_utils import ReportTypes, ReportMeta
-from ppr_api.callback.auth.token_service import GoogleStorageTokenService
+from ppr_api.reports.v2.report_utils import ReportMeta, ReportTypes
 
 
 SINGLE_URI = '/forms/chromium/convert/html'
@@ -72,7 +71,7 @@ class Report:  # pylint: disable=too-few-public-methods
         token = GoogleStorageTokenService.get_report_api_token()
         if token:
             headers['Authorization'] = 'Bearer {}'.format(token)
-        response = requests.post(url=url, headers=headers, data=meta_data, files=files)
+        response = requests.post(url=url, headers=headers, data=meta_data, files=files, timeout=1800.0)
         current_app.logger.debug('Account {0} report type {1} response status: {2}.'
                                  .format(self._account_id, self._report_key, response.status_code))
         if response.status_code != HTTPStatus.OK:
@@ -99,7 +98,7 @@ class Report:  # pylint: disable=too-few-public-methods
             headers['Authorization'] = 'Bearer {}'.format(token)
 
         files = report_utils.get_report_files(data, self._report_key, False, False)
-        response_reg = requests.post(url=url, headers=headers, data=meta_data, files=files)
+        response_reg = requests.post(url=url, headers=headers, data=meta_data, files=files, timeout=1800.0)
         current_app.logger.debug('Account {0} report type {1} response status: {2}.'
                                  .format(self._account_id, self._report_key, response_reg.status_code))
         if response_reg.status_code != HTTPStatus.OK:
@@ -115,7 +114,7 @@ class Report:  # pylint: disable=too-few-public-methods
                                  .format(self._account_id, self._report_key, url))
         files = report_utils.get_report_files(data_final, self._report_key, False, False)
         current_app.logger.info('Search report regenerating with TOC page numbers set.')
-        response = requests.post(url=url, headers=headers, data=meta_data, files=files)
+        response = requests.post(url=url, headers=headers, data=meta_data, files=files, timeout=1800.0)
         current_app.logger.info('Search report regeneration with TOC page numbers completed.')
         if response.status_code != HTTPStatus.OK:
             content = ResourceErrorCodes.REPORT_ERR + ': ' + response.content.decode('ascii')
@@ -199,7 +198,7 @@ class Report:  # pylint: disable=too-few-public-methods
         token = GoogleStorageTokenService.get_report_api_token()
         if token:
             headers['Authorization'] = 'Bearer {}'.format(token)
-        response_cover = requests.post(url=url, headers=headers, data=meta_data, files=files)
+        response_cover = requests.post(url=url, headers=headers, data=meta_data, files=files, timeout=1800.0)
         current_app.logger.debug('Account {0} report type {1} response status: {2}.'
                                  .format(self._account_id, self._report_key, response_cover.status_code))
         if response_cover.status_code != HTTPStatus.OK:
@@ -216,7 +215,7 @@ class Report:  # pylint: disable=too-few-public-methods
                                  .format(self._account_id, self._report_key, url))
         meta_data = report_utils.get_report_meta_data(self._report_key)
         files = report_utils.get_report_files(data, self._report_key, True)
-        response_reg = requests.post(url=url, headers=headers, data=meta_data, files=files)
+        response_reg = requests.post(url=url, headers=headers, data=meta_data, files=files, timeout=1800.0)
         current_app.logger.debug('Account {0} report type {1} response status: {2}.'
                                  .format(self._account_id, self._report_key, response_reg.status_code))
         if response_reg.status_code != HTTPStatus.OK:
@@ -230,7 +229,7 @@ class Report:  # pylint: disable=too-few-public-methods
             'pdf1.pdf': response_cover.content,
             'pdf2.pdf': response_reg.content
         }
-        response = requests.post(url=url, headers=headers, files=files)
+        response = requests.post(url=url, headers=headers, files=files, timeout=1800.0)
         current_app.logger.debug('Merge cover and registration reports response status: {0}.'
                                  .format(response.status_code))
         if response.status_code != HTTPStatus.OK:
@@ -287,7 +286,7 @@ class Report:  # pylint: disable=too-few-public-methods
         """Load from the local file system the template matching the report type."""
         try:
             template_path = current_app.config.get('REPORT_TEMPLATE_PATH')
-            template_code = Path(f'{template_path}/{self._get_template_filename()}').read_text()
+            template_code = Path(f'{template_path}/{self._get_template_filename()}').read_text(encoding='UTF-8')
             # substitute template parts
             template_code = self._substitute_template_parts(template_code)
         except Exception as err:  # noqa: B902; just logging
@@ -352,12 +351,13 @@ class Report:  # pylint: disable=too-few-public-methods
         # substitute template parts - marked up by [[filename]]
         for template_part in template_parts:
             if template_code.find('[[{}.html]]'.format(template_part)) >= 0:
-                template_part_code = Path(f'{template_path}/template-parts/{template_part}.html').read_text()
+                template_part_code = \
+                    Path(f'{template_path}/template-parts/{template_part}.html').read_text(encoding='UTF-8')
                 for template_part_nested in template_parts:
                     template_reference = '[[{}.html]]'.format(template_part_nested)
                     if template_part_code.find(template_reference) >= 0:
                         path = Path(f'{template_path}/template-parts/{template_part_nested}.html')
-                        template_nested_code = path.read_text()
+                        template_nested_code = path.read_text(encoding='UTF-8')
                         template_part_code = template_part_code.replace(template_reference, template_nested_code)
                 template_code = template_code.replace('[[{}.html]]'.format(template_part), template_part_code)
 
@@ -675,14 +675,14 @@ class Report:  # pylint: disable=too-few-public-methods
                     else:
                         add_party['address_change'] = True
                     break
-                elif 'amendPartyId' not in add_party:
+                if 'amendPartyId' not in add_party:
                     if add_party['address'] == delete_party['address']:
                         if 'businessName' in add_party and 'businessName' in delete_party and \
                                 add_party['businessName'] != delete_party['businessName']:
                             add_party['name_change'] = True
                             delete_party['edit'] = True
                             break
-                        elif 'personName' in add_party and 'personName' in delete_party and \
+                        if 'personName' in add_party and 'personName' in delete_party and \
                                 add_party['personName'] != delete_party['personName']:
                             add_party['name_change'] = True
                             delete_party['edit'] = True
@@ -838,7 +838,7 @@ class Report:  # pylint: disable=too-few-public-methods
 
         return address
 
-    def _set_meta_info(self):
+    def _set_meta_info(self):  # pylint: disable=too-many-branches; only 1 more.
         """Identify environment in report if non-production."""
         self._report_data['environment'] = f'{self._get_environment()}'.lstrip()
         self._report_data['meta_account_id'] = self._account_id
