+++ ppr-api/src/ppr_api/resources/v1/callbacks.py
@@ -15,112 +15,107 @@
 import json
 from http import HTTPStatus
 
-from flask import request, current_app, jsonify
-from flask_restx import Namespace, Resource, cors
-
+from flask import Blueprint, current_app, jsonify, request
+from flask_cors import cross_origin
+from ppr_api.callback.document_storage.storage_service import DocumentTypes, GoogleStorageService
+from ppr_api.callback.utils.exceptions import ReportDataException, ReportException, StorageException
 from ppr_api.exceptions import DatabaseException
 from ppr_api.models import MailReport
 from ppr_api.models import utils as model_utils
-from ppr_api.resources import utils as resource_utils
-from ppr_api.utils.util import cors_preflight
 from ppr_api.reports import get_callback_pdf
 from ppr_api.reports.v2.report_utils import ReportTypes
-from ppr_api.callback.utils.exceptions import ReportDataException, ReportException, StorageException
-from ppr_api.callback.document_storage.storage_service import DocumentTypes, GoogleStorageService
+from ppr_api.resources import utils as resource_utils
 
 
-API = Namespace('callbacks', description='Endpoints for processing aysnchronous callback requests.')
+bp = Blueprint('CALLBACKS1', __name__,  # pylint: disable=invalid-name
+               url_prefix='/api/v1/callbacks')
 START_TS_PARAM = 'startDateTime'
 END_TS_PARAM = 'endDateTime'
 
 
-@cors_preflight('GET,POST,OPTIONS')
-@API.route('/mail-report', methods=['GET', 'POST', 'OPTIONS'])
-class MailReportResource(Resource):
-    """Resource to handle a verification report mail requests for a callback events."""
+@bp.route('/mail-report', methods=['POST', 'OPTIONS'])
+@cross_origin(origin='*')
+def post_mail_report():  # pylint: disable=too-many-return-statements
+    """Registration id and party id are in the payload."""
+    request_json = request.get_json(silent=True)
+    registration_id: int = request_json.get('registrationId', -1)
+    party_id: int = request_json.get('partyId', -1)
+    try:
+        # Authenticate with request api key
+        if not resource_utils.valid_api_key(request):
+            return resource_utils.unauthorized_error_response('Verification report callback')
 
-    @staticmethod
-    @cors.crossdomain(origin='*')
-    def post():
-        """Registration id and party id are in the payload."""
-        request_json = request.get_json(silent=True)
-        registration_id: int = request_json.get('registrationId', -1)
-        party_id: int = request_json.get('partyId', -1)
-        try:
-            # Authenticate with request api key
-            if not resource_utils.valid_api_key(request):
-                return resource_utils.unauthorized_error_response('Verification report callback')
+        message: str = None
+        status = HTTPStatus.BAD_REQUEST
+        if registration_id < 0:
+            message = 'Mail verification statement no registration ID.'
+        elif party_id < 0:
+            message = 'Mail verification statement no party ID.'
+        else:
+            # If no mail report record we're done.
+            mail_report: MailReport = MailReport.find_by_registration_party_id(registration_id, party_id)
+            if not mail_report:
+                message = 'No mail report data found for the registration id, party id.'
+                status = HTTPStatus.NOT_FOUND
 
-            message: str = None
-            status = HTTPStatus.BAD_REQUEST
-            if registration_id < 0:
-                message = 'Mail verification statement no registration ID.'
-            elif party_id < 0:
-                message = 'Mail verification statement no party ID.'
-            else:
-                # If no mail report record we're done.
-                mail_report: MailReport = MailReport.find_by_registration_party_id(registration_id, party_id)
-                if not mail_report:
-                    message = 'No mail report data found for the registration id, party id.'
-                    status = HTTPStatus.NOT_FOUND
+        if message:
+            return resource_utils.error_response(status, message)
 
-            if message:
-                return resource_utils.error_response(status, message)
+        if mail_report.doc_storage_url:
+            current_app.logger.debug('Report {mail_report.doc_storage_url} already exists.')
+            return {}, HTTPStatus.OK, {'Content-Type': 'application/json'}
+        # Generate the report, upload it to cloud storage, update the status.
+        return generate_mail_callback_report(mail_report)
+    except DatabaseException as db_exception:
+        return resource_utils.db_exception_response(db_exception,
+                                                    None,
+                                                    'POST callback mail verification report DB error.')
+    except Exception as default_err:  # noqa: B902; return nicer default error
+        return mail_callback_error(mail_report,
+                                   HTTPStatus.INTERNAL_SERVER_ERROR,
+                                   'Callback default error: ' + str(default_err))
 
-            if mail_report.doc_storage_url:
-                current_app.logger.debug('Report {mail_report.doc_storage_url} already exists.')
-                return {}, HTTPStatus.OK, {'Content-Type': 'application/json'}
-            # Generate the report, upload it to cloud storage, update the status.
-            return generate_mail_callback_report(mail_report)
-        except DatabaseException as db_exception:
-            return resource_utils.db_exception_response(db_exception,
-                                                        None,
-                                                        'POST callback mail verification report DB error.')
-        except Exception as default_err:  # noqa: B902; return nicer default error
-            return mail_callback_error(mail_report,
-                                       HTTPStatus.INTERNAL_SERVER_ERROR,
-                                       'Callback default error: ' + str(default_err))
 
-    @staticmethod
-    @cors.crossdomain(origin='*')
-    def get():
-        """Fetch recent event storage names by request parameter startDateTime and optional endDateTime."""
-        start_ts = request.args.get(START_TS_PARAM, None)
-        end_ts = request.args.get(END_TS_PARAM, None)
-        try:
-            # Authenticate with request api key
-            if not resource_utils.valid_api_key(request):
-                return resource_utils.unauthorized_error_response('Verification report callback')
+@bp.route('/mail-report', methods=['GET', 'OPTIONS'])
+@cross_origin(origin='*')
+def get_mail_list():
+    """Fetch recent event storage names by request parameter startDateTime and optional endDateTime."""
+    start_ts = request.args.get(START_TS_PARAM, None)
+    end_ts = request.args.get(END_TS_PARAM, None)
+    try:
+        # Authenticate with request api key
+        if not resource_utils.valid_api_key(request):
+            return resource_utils.unauthorized_error_response('Verification report callback')
 
-            message: str = None
-            status = HTTPStatus.BAD_REQUEST
-            start = None
-            end = None
-            if not start_ts:
-                message = 'Request parameter startDateTime is required.'
-            else:
-                try:
-                    start = model_utils.ts_from_iso_format(start_ts)
-                except Exception:
-                    message = 'Request parameter startDateTime value is invalid'
-            if end_ts:
-                try:
-                    end = model_utils.ts_from_iso_format(end_ts)
-                except Exception:
-                    message = 'Request parameter endDateTime value is invalid'
-            if end and start and end <= start:
-                message = 'Request timestamp range is invalid'
+        message: str = None
+        status = HTTPStatus.BAD_REQUEST
+        start = None
+        end = None
+        if not start_ts:
+            message = 'Request parameter startDateTime is required.'
+        else:
+            try:
+                start = model_utils.ts_from_iso_format(start_ts)
+            except Exception:  # noqa: B902; validation error.
+                message = 'Request parameter startDateTime value is invalid'
+        if end_ts:
+            try:
+                end = model_utils.ts_from_iso_format(end_ts)
+            except Exception:  # noqa: B902; validation error
+                message = 'Request parameter endDateTime value is invalid'
+        if end and start and end <= start:
+            message = 'Request timestamp range is invalid'
 
-            if message:
-                return resource_utils.error_response(status, message)
-            results = MailReport.find_list_by_timestamp(start, end)
-            return jsonify(results), HTTPStatus.OK
-        except DatabaseException as db_exception:
-            return resource_utils.db_exception_response(db_exception,
-                                                        None,
-                                                        'GET callback mail reports list DB error.')
-        except Exception as default_err:  # noqa: B902; return nicer default error
-            return resource_utils.default_exception_response(default_err)
+        if message:
+            return resource_utils.error_response(status, message)
+        results = MailReport.find_list_by_timestamp(start, end)
+        return jsonify(results), HTTPStatus.OK
+    except DatabaseException as db_exception:
+        return resource_utils.db_exception_response(db_exception,
+                                                    None,
+                                                    'GET callback mail reports list DB error.')
+    except Exception as default_err:  # noqa: B902; return nicer default error
+        return resource_utils.default_exception_response(default_err)
 
 
 def mail_callback_error(mail_report: MailReport, status_code: int = 500, message: str = None):
