+++ ppr-api/src/ppr_api/services/payment/client/__init__.py
@@ -22,7 +22,6 @@ from functools import wraps
 
 import requests
 from flask import current_app
-
 from ppr_api.services.payment import TransactionTypes
 
 
@@ -244,14 +243,16 @@ class BaseClient:
                     url,
                     params=None,
                     json=data,
-                    headers=headers
+                    headers=headers,
+                    timeout=3.0
                 )
             else:
                 response = requests.request(
                     method.value,
                     url,
                     params=None,
-                    headers=headers
+                    headers=headers,
+                    timeout=3.0
                 )
 
             if response is not None:
@@ -478,7 +479,8 @@ class SBCPaymentClient(BaseClient):
                 oidc_token_url,
                 data=data,
                 params=None,
-                headers=headers
+                headers=headers,
+                timeout=3.0
             )
 
             if not response or not response.ok:
